pyproject.toml
[project]
name = "db-assisst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.11",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
